requirements.txt
transformers==4.57.3
torch==2.9.1
accelerate==1.12.0
sentencepiece==0.2.1
srt
pyyaml
